# legacy/brain/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for Arthemis Brain.",
        "title": "Arthemis Brain API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/location/create": {
            "post": {
                "description": "Creates multiple locations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Create Locations",
                "parameters": [
                    {
                        "description": "List of location details",
                        "name": "locations",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Location"
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "IDs of created locations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/location/delete/{id}": {
            "delete": {
                "description": "Deletes a Location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Delete Location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location removed successfully",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/location/update/{id}": {
            "put": {
                "description": "Updates a location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Update Location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Location details to update",
                        "name": "project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location updated successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/location/{id}": {
            "get": {
                "description": "Retrieves a location",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "location"
                ],
                "summary": "Get Location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Location ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location retrtieved successfully",
                        "schema": {
                            "$ref": "#/definitions/models.Location"
                        }
                    },
                    "400": {
                        "description": "ID not informed",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Location not found",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "models.Indicator": {
            "type": "object",
            "properties": {
                "activityID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "justification": {
                    "type": "string"
                },
                "locationID": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "observationMethod": {
                    "type": "string"
                },
                "observations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Observation"
                    }
                },
                "unit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "valueBaseline": {
                    "type": "number"
                },
                "valueReference": {
                    "type": "number"
                }
            }
        },
        "models.JSONB": {
            "type": "object",
            "additionalProperties": {}
        },
        "models.Location": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "ecosystem": {
                    "type": "string"
                },
                "extent": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "indicators": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Indicator"
                    }
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "projectID": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Observation": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "indicatorID": {
                    "type": "integer"
                },
                "position": {
                    "$ref": "#/definitions/models.JSONB"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Tarefa não encontrada"
                },
                "message": {
                    "type": "string",
                    "example": "ID inválido fornecido"
                }
            }
        }
    }
}